requirements.txt
behave
selenium
webdriver-manager
python-dotenv
requests
allure-behave
pytest
pytest-playwright
playwright
httpx
